Moves Setup Wizard to it owns chunk
Setup wizard is only needed on native devices such Android and iOS,
with that in mind, this commit moves the whole setup wizard to it owns
chunks, so it can be loaded on demand and isn't automatically loaded in
desktop or web app.

Closes #6844
