+++ src/native/main/wizard/SetupWizard.ts
@@ -1,27 +1,23 @@
-import { createWizardDialog, wizardPageWrapper } from "../base/WizardDialog.js"
+import { createWizardDialog, wizardPageWrapper } from "../../../gui/base/WizardDialog.js"
 import { defer } from "@tutao/tutanota-utils"
 import { SetupCongratulationsPage, SetupCongratulationsPageAttrs } from "./setupwizardpages/SetupCongraulationsPage.js"
-import { DeviceConfig, deviceConfig } from "../../misc/DeviceConfig.js"
+import { DeviceConfig } from "../../../misc/DeviceConfig.js"
 import m from "mithril"
-import { isAndroidApp, isApp } from "../../api/common/Env.js"
 import { SetupNotificationsPage, SetupNotificationsPageAttrs } from "./setupwizardpages/SetupNotificationsPage.js"
-import { BannerButton } from "../base/buttons/BannerButton.js"
-import { theme } from "../theme.js"
-import { ClickHandler } from "../base/GuiUtils.js"
-import { PermissionType } from "../../native/common/generatedipc/PermissionType.js"
-import { locator } from "../../api/main/MainLocator.js"
-import { PermissionError } from "../../api/common/error/PermissionError.js"
-import { Dialog, DialogType } from "../base/Dialog.js"
-import { TranslationKey } from "../../misc/LanguageViewModel.js"
+import { BannerButton } from "../../../gui/base/buttons/BannerButton.js"
+import { theme } from "../../../gui/theme.js"
+import { ClickHandler } from "../../../gui/base/GuiUtils.js"
+import { DialogType } from "../../../gui/base/Dialog.js"
+import { TranslationKey } from "../../../misc/LanguageViewModel.js"
 import { SetupThemePage, SetupThemePageAttrs } from "./setupwizardpages/SetupThemePage.js"
 import { SetupContactsPage, SetupContactsPageAttrs } from "./setupwizardpages/SetupContactsPage.js"
 import { SetupLockPage, SetupLockPageAttrs } from "./setupwizardpages/SetupLockPage.js"
-import { SystemPermissionHandler } from "../../native/main/SystemPermissionHandler.js"
-import { WebMobileFacade } from "../../native/main/WebMobileFacade.js"
-import { ContactImporter } from "../../contacts/ContactImporter.js"
-import { MobileSystemFacade } from "../../native/common/generatedipc/MobileSystemFacade.js"
-import { CredentialsProvider } from "../../misc/credentials/CredentialsProvider.js"
-import { NativeContactsSyncManager } from "../../contacts/model/NativeContactsSyncManager.js"
+import { SystemPermissionHandler } from "../SystemPermissionHandler.js"
+import { WebMobileFacade } from "../WebMobileFacade.js"
+import { ContactImporter } from "../../../contacts/ContactImporter.js"
+import { MobileSystemFacade } from "../../common/generatedipc/MobileSystemFacade.js"
+import { CredentialsProvider } from "../../../misc/credentials/CredentialsProvider.js"
+import { NativeContactsSyncManager } from "../../../contacts/model/NativeContactsSyncManager.js"
 
 export function renderPermissionButton(permissionName: TranslationKey, isPermissionGranted: boolean, onclick: ClickHandler) {
 	return renderBannerButton(isPermissionGranted ? "granted_msg" : permissionName, onclick, isPermissionGranted)
