+++ src/native/main/wizard/setupwizardpages/SetupLockPage.ts
@@ -1,15 +1,15 @@
 import m, { Children, Component, Vnode } from "mithril"
-import { WizardPageAttrs } from "../../base/WizardDialog.js"
-import { lang } from "../../../misc/LanguageViewModel.js"
+import { WizardPageAttrs } from "../../../../gui/base/WizardDialog.js"
+import { lang } from "../../../../misc/LanguageViewModel.js"
 import { SetupPageLayout } from "./SetupPageLayout.js"
-import { Dialog } from "../../base/Dialog.js"
-import { DEFAULT_CREDENTIAL_ENCRYPTION_MODE, SelectCredentialsEncryptionModeView } from "../SelectCredentialsEncryptionModeDialog.js"
-import { CredentialAuthenticationError } from "../../../api/common/error/CredentialAuthenticationError.js"
-import { KeyPermanentlyInvalidatedError } from "../../../api/common/error/KeyPermanentlyInvalidatedError.js"
-import { windowFacade } from "../../../misc/WindowFacade.js"
-import { CancelledError } from "../../../api/common/error/CancelledError.js"
-import { CredentialsProvider } from "../../../misc/credentials/CredentialsProvider.js"
-import { CredentialEncryptionMode } from "../../../misc/credentials/CredentialEncryptionMode.js"
+import { Dialog } from "../../../../gui/base/Dialog.js"
+import { DEFAULT_CREDENTIAL_ENCRYPTION_MODE, SelectCredentialsEncryptionModeView } from "../../../../gui/dialogs/SelectCredentialsEncryptionModeDialog.js"
+import { CredentialAuthenticationError } from "../../../../api/common/error/CredentialAuthenticationError.js"
+import { KeyPermanentlyInvalidatedError } from "../../../../api/common/error/KeyPermanentlyInvalidatedError.js"
+import { windowFacade } from "../../../../misc/WindowFacade.js"
+import { CancelledError } from "../../../../api/common/error/CancelledError.js"
+import { CredentialsProvider } from "../../../../misc/credentials/CredentialsProvider.js"
+import { CredentialEncryptionMode } from "../../../../misc/credentials/CredentialEncryptionMode.js"
 
 export class SetupLockPage implements Component<SetupLockPageAttrs> {
 	view({ attrs }: Vnode<SetupLockPageAttrs>): Children {
